Sync with upstream

+++ xmrstak/backend/cpu/crypto/hash.h
@@ -1,5 +1,12 @@
 #pragma once
 
+#include <stdint.h>
+
 typedef unsigned char BitSequence;
-typedef unsigned long long DataLength;
-typedef enum {SUCCESS = 0, FAIL = 1, BAD_HASHLEN = 2} HashReturn;
+typedef uint32_t DataLength;
+typedef enum
+{
+	SUCCESS = 0,
+	FAIL = 1,
+	BAD_HASHLEN = 2
+} HashReturn;
